+++ example/src/App.tsx
@@ -39,11 +39,10 @@ const App: React.FC = () => {
       packageName: 'io.ionic.starter',
       certificateHashes: ['AKoRuyLMM91E7lX/Zqp3u4jMmd0A7hH/Iqozu0TMVd0='],
       // supportedAlternativeStores: ['storeOne', 'storeTwo'],
-      // @deprecated: blacklistedHashes, blacklistedPackageNames, suspiciousPermissions, whitelistedInstallationSources
-      malwareConfig: {
-        blacklistedHashes: ['FgvSehLMM91E7lX/Zqp3u4jMmd0A7hH/Iqozu0TMVd0u'],
-        blacklistedPackageNames: ['io.ionic.starter'],
-        suspiciousPermissions: [
+      suspiciousAppDetectionConfig: {
+        packageNames: ['io.ionic.starter'],
+        hashes: ['FgvSehLMM91E7lX/Zqp3u4jMmd0A7hH/Iqozu0TMVd0u'],
+        requestedPermissions: [
           [
             'android.permission.INTERNET',
             'android.permission.ACCESS_COARSE_LOCATION',
@@ -51,7 +50,10 @@ const App: React.FC = () => {
           ['android.permission.BLUETOOTH'],
           ['android.permission.BATTERY_STATS'],
         ],
-        whitelistedInstallationSources: ['com.apkpure.aegon'],
+        malwareScanScope: {
+          scanScope: 'SIDELOADED_ONLY',
+          trustedInstallSources: ['com.apkpure.aegon'],
+        },
       },
     },
     iosConfig: {