+++ src/types/types.ts
@@ -28,6 +28,7 @@ export type TalsecAndroidConfig = {
   certificateHashes: string[];
   supportedAlternativeStores?: string[];
   malwareConfig?: TalsecMalwareConfig;
+  suspiciousAppDetectionConfig?: SuspiciousAppDetectionConfig;
 };
 
 export type TalsecIosConfig = {
@@ -36,14 +37,37 @@ export type TalsecIosConfig = {
 };
 
 export type TalsecMalwareConfig = {
-  /** @deprecated */
+  /** @deprecated Use SuspiciousAppDetectionConfig instead */
   blacklistedHashes?: string[];
-  /** @deprecated */
+  /** @deprecated Use SuspiciousAppDetectionConfig instead */
   blacklistedPackageNames?: string[];
-  /** @deprecated */
+  /** @deprecated Use SuspiciousAppDetectionConfig instead */
   suspiciousPermissions?: string[][];
-  /** @deprecated */
+  /** @deprecated Use SuspiciousAppDetectionConfig instead */
   whitelistedInstallationSources?: string[];
+};
+
+export type ScopeType =
+  | 'SIDELOADED_ONLY'
+  | 'SIDELOADED_AND_SYSTEM_EXCLUDE_OEM'
+  | 'SIDELOADED_AND_OEM'
+  | 'SIDELOADED_AND_SYSTEM_AND_OEM'
+  | 'ALL';
+
+export type ReasonMode = 'ALL' | 'HIGHEST_CONFIDENCE';
+
+export type MalwareScanScope = {
+  scanScope: ScopeType;
+  trustedInstallSources?: string[];
+};
+
+export type SuspiciousAppDetectionConfig = {
+  packageNames?: string[];
+  hashes?: string[];
+  requestedPermissions?: string[][];
+  grantedPermissions?: string[][];
+  malwareScanScope?: MalwareScanScope;
+  reasonMode?: ReasonMode;
 };
 
 export type SuspiciousAppInfo = {